Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaDevices-enumerateDevices.https-expected.txt

Description

This test checks for the presence of the navigator.mediaDevices.enumerateDevices() method.


PASS mediaDevices.enumerateDevices() is present and working - before capture
PASS mediaDevices.enumerateDevices() is present and working - after capture
FAIL InputDeviceInfo is supported promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: InputDeviceInfo"
